// frontend/src/app/models/comment.model.ts
import { Profile } from './profile.model';

export interface Comment {
  id: number;
  createdAt: string;
  updatedAt: string;
  body: string;
  author: Profile;
}

export interface CommentResponse {
  comment: Comment;
}

export interface CommentsResponse {
  comments: Comment[];
}

export interface CreateCommentRequest {
  comment: {
    body: string;
  };
}